fix: fixed unit test
