--- converter/tests/integration_test.rs
@@ -1,17 +1,0 @@
-use crate::converter::{celsius_to_fahrenheit, fahrenheit_to_celsius};
-
-#[cfg(test)]
-mod tests {
-    fn test_fahrenheit_to_celsius() {
-        assert_eq!(fahrenheit_to_celsius(32.0), 0.0);
-        assert_eq!(fahrenheit_to_celsius(212.0), 100.0);
-        assert_eq!(fahrenheit_to_celsius(68.0), 20.0);
-    }
-
-    #[test]
-    fn test_celsius_to_fahrenheit() {
-        assert_eq!(celsius_to_fahrenheit(0.0), 32.0);
-        assert_eq!(celsius_to_fahrenheit(100.0), 212.0);
-        assert_eq!(celsius_to_fahrenheit(20.0), 68.0);
-    }
-}
